#341 Refactor: split fetchAPI to multiple file like product service, status service
