+++ react/practice-two/src/services/index.ts
@@ -1,1 +1,3 @@
-export * from './fetchAPI';
+export * from './productsService';
+export * from './statusesService';
+export * from './typesService';
