+++ react/practice-two/src/services/productsService.ts
@@ -1,7 +1,8 @@
+// Constants
 import { URL_API } from '@constants';
-import { CustomErrors, customErrors } from '@helpers';
 
-// Products
+// Helpers
+import { CustomErrors, customErrors } from '@helpers';
 
 /**
  * @description function get all products
@@ -51,22 +52,22 @@ const deleteProduct = async <T>(id: string): Promise<T | CustomErrors> => {
  * @description function update product which is selected
  *
  * @param {String} id is id of product
- * @param {Object} product which is selected
+ * @param {Object} productUpdate is information about product, which is selected
  *
  * @return {Object} product
  */
-const updateProduct = async <T>(id: string, product: T): Promise<T | CustomErrors> => {
+const updateProduct = async <T>(id: string, productUpdate: T): Promise<T | CustomErrors> => {
   try {
     const options = {
       method: 'PATCH',
-      body: JSON.stringify(product),
+      body: JSON.stringify(productUpdate),
       headers: {
         'Content-Type': 'application/json',
       },
     };
     const response = await fetch(`${URL_API.BASE_URL}${URL_API.PRODUCTS}/${id}`, options);
-    const products: T = await response.json();
-    const productItem = customErrors(response, products);
+    const product: T = await response.json();
+    const productItem = customErrors(response, product);
 
     return productItem;
   } catch (error) {
@@ -74,41 +75,4 @@ const updateProduct = async <T>(id: string, product: T): Promise<T | CustomError
   }
 };
 
-// Types
-
-/**
- * @description function get all types
- *
- * @returns {Array} list item
- */
-const getTypes = async <T>(): Promise<T[] | CustomErrors> => {
-  try {
-    const response = await fetch(`${URL_API.BASE_URL}${URL_API.TYPES}`);
-    const types: T[] = await response.json();
-    const listTypes = customErrors(response, types);
-
-    return listTypes;
-  } catch (error) {
-    return error as CustomErrors;
-  }
-};
-
-// Statuses
-
-/**
- * @description function get all statuses
- *
- * @returns {Array} list statuses
- */
-const getStatuses = async <T>(): Promise<T[] | CustomErrors> => {
-  try {
-    const response = await fetch(`${URL_API.BASE_URL}${URL_API.STATUSES}`);
-    const status: T[] = await response.json();
-    const listStatuses = customErrors(response, status);
-
-    return listStatuses;
-  } catch (error) {
-    return error as CustomErrors;
-  }
-};
-export { getProductsByParam, deleteProduct, updateProduct, getTypes, getStatuses };
+export { updateProduct, getProductsByParam, deleteProduct };
